Split the file cache into cache and garbage collector
WIP #34
