+++ test/Cache/File/CacheTest.php
@@ -2,20 +2,7 @@
 
 namespace Kibo\Phast\Cache\File;
 
-use Kibo\Phast\Common\ObjectifiedFunctions;
-use PHPUnit\Framework\TestCase;
-
-class CacheTest extends TestCase {
-
-    /**
-     * @var array
-     */
-    private $config;
-
-    /**
-     * @var ObjectifiedFunctions
-     */
-    private $functions;
+class CacheTest extends CacheTestCase {
 
     private $nameSpace = 'test';
 
@@ -26,18 +13,6 @@ class CacheTest extends TestCase {
 
     public function setUp() {
         parent::setUp();
-        $this->config = [
-            'cacheRoot' => sys_get_temp_dir() . '/test-cache-dir-' . posix_geteuid(),
-            'garbageCollection' => [
-                'probability' => 0,
-                'maxItems' => 100,
-                'maxAge' => 20
-            ]
-        ];
-        $this->rmDir($this->config['cacheRoot']);
-        $this->functions = new ObjectifiedFunctions();
-        $this->functions->register_shutdown_function = function () {};
-        $this->functions->error_log = function () {};
         $this->rebuildCache();
     }
 
@@ -153,129 +128,6 @@ class CacheTest extends TestCase {
         $this->cache->get('asd', function () {});
         $this->assertTrue($touched);
 
-    }
-
-    public function testGarbageCollectionRunCalculation() {
-        $actualMin = $actualMax = null;
-        $this->functions->mt_rand = function ($min, $max) use (&$actualMin, &$actualMax) {
-            $actualMin = $min;
-            $actualMax = $max;
-        };
-        $this->functions->file_exists = function () {
-            return true;
-        };
-        $this->config['garbageCollection']['probability'] = 0.5;
-        $this->rebuildCache();
-        $this->assertEquals(1, $actualMin);
-        $this->assertEquals(2, $actualMax);
-    }
-
-    public function testNotExplodingWhenGarbageCollectingOnMissingCacheRoot() {
-        $this->config['garbageCollection']['probability'] = 1;
-        $this->rebuildCache();
-    }
-
-    public function testGarbageCollection() {
-        $results = $this->executeCacheTest();
-
-        $dir1 = $results['dir1'];
-        $this->assertCount(2, $dir1);
-        $this->assertContains('item1', $dir1);
-        $this->assertContains('a-dir', $dir1);
-        $this->assertNotContains('item2', $dir1);
-        $this->assertNotContains('item3', $dir1);
-        $this->assertNotContains('item4', $dir1);
-
-        $dir2 = $results['dir2'];
-        $this->assertCount(3, $dir2);
-        $this->assertContains('item4', $dir2);
-        $this->assertContains('item5', $dir2);
-        $this->assertContains('a-dir', $dir2);
-        $this->assertNotContains('item6', $dir2);
-        $this->assertNotContains('item7', $dir2);
-    }
-
-    public function testGarbageCollectionLimit() {
-        $this->config['garbageCollection']['maxItems'] = 2;
-        $result = $this->executeCacheTest();
-
-        $totalCount = count($result['dir1']) + count($result['dir2']);
-        $this->assertEquals(7, $totalCount);
-
-        $dir1 = $result['dir1'];
-        $this->assertContains('item1', $dir1);
-        $this->assertContains('a-dir', $dir1);
-
-        $dir2 = $result['dir2'];
-        $this->assertContains('item4', $dir2);
-        $this->assertContains('item5', $dir2);
-        $this->assertContains('a-dir', $dir2);
-    }
-
-    private function executeCacheTest() {
-        $this->config['garbageCollection']['probability'] = 1;
-        $this->config['garbageCollection']['maxAge'] = 60;
-        $items = [
-            'dir1' => [
-                'item1' => 30,
-                'item2' => 100,
-                'item3' => 200
-            ],
-            'dir2' => [
-                'item4' => 40,
-                'item5' => 45,
-                'item6' => 90,
-                'item7' => 70
-            ]
-        ];
-
-        foreach ($items as $dir => $files) {
-            $path = $this->config['cacheRoot'] . '/' . $dir;
-            mkdir($path, 0700, true);
-            mkdir($path . '/a-dir', 0700);
-            foreach ($files as $file => $time) {
-                $filename = $path . '/' . $file;
-                touch($filename, time() - $time);
-            }
-        }
-
-        $this->rebuildCache();
-
-        $getDirItems = function ($path) {
-            return array_filter(scandir($path), function ($item) {
-                return !in_array($item, ['.', '..']);
-            });
-        };
-        $dirs = $getDirItems($this->config['cacheRoot']);
-        $results = [];
-        foreach ($dirs as $dir) {
-            $path = $this->config['cacheRoot'] . '/' . $dir;
-            $results[$dir] = $getDirItems($path);
-        }
-
-        $this->assertCount(2, $results);
-        $this->assertArrayHasKey('dir1', $results);
-        $this->assertArrayHasKey('dir2', $results);
-
-        return $results;
-    }
-
-    private function rmDir($dir) {
-        if (!file_exists($dir)) {
-            return;
-        }
-        foreach (scandir($dir) as $item) {
-            if ($item == '.' || $item == '..') {
-                continue;
-            }
-            $fullName = $dir . '/' . $item;
-            if (is_dir($fullName)) {
-                $this->rmDir($fullName);
-            } else {
-                unlink($fullName);
-            }
-        }
-        rmdir($dir);
     }
 
     private function rebuildCache() {
